+++ br/pkg/checksum/executor_test.go
@@ -8,10 +8,10 @@ import (
 	"testing"
 
 	"github.com/pingcap/failpoint"
-	"github.com/pingcap/tidb/br/pkg/backup"
 	"github.com/pingcap/tidb/br/pkg/checksum"
 	"github.com/pingcap/tidb/br/pkg/metautil"
 	"github.com/pingcap/tidb/br/pkg/mock"
+	"github.com/pingcap/tidb/pkg/distsql"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
@@ -132,7 +132,7 @@ func TestChecksum(t *testing.T) {
 	require.NoError(t, exe3.Each(func(req *kv.Request) error {
 		if first {
 			first = false
-			ranges, err := backup.BuildTableRanges(tableInfo3)
+			ranges, err := distsql.BuildTableRanges(tableInfo3)
 			require.NoError(t, err)
 			require.Equalf(t, ranges[:1], req.KeyRanges.FirstPartitionRange(), "%v", req.KeyRanges.FirstPartitionRange())
 		}
