+++ br/pkg/lightning/common/util_test.go
@@ -302,3 +302,35 @@ CREATE TABLE multi_indexes (
 		require.Equal(t, tt.multiSQLs, multiSQLs)
 	}
 }
+
+func TestHideSensitive(t *testing.T) {
+	strs := []struct {
+		old string
+		new string
+	}{
+		{
+			`host = "127.0.0.1"\n  user = "root"\n  password = "/Q7B9DizNLLTTfiZHv9WoEAKamfpIUs="\n  port = 3306\n`,
+			`host = "127.0.0.1"\n  user = "root"\n  password = ******\n  port = 3306\n`,
+		},
+		{
+			`host = "127.0.0.1"\n  user = "root"\n  password = ""\n  port = 3306\n`,
+			`host = "127.0.0.1"\n  user = "root"\n  password = ******\n  port = 3306\n`,
+		},
+		{
+			`host = "127.0.0.1"\n  user = "root"\n  password= "/Q7B9DizNLLTTfiZHv9WoEAKamfpIUs="\n  port = 3306\n`,
+			`host = "127.0.0.1"\n  user = "root"\n  password= ******\n  port = 3306\n`,
+		},
+		{
+			`host = "127.0.0.1"\n  user = "root"\n  password =""\n  port = 3306\n`,
+			`host = "127.0.0.1"\n  user = "root"\n  password =******\n  port = 3306\n`,
+		},
+		{
+			`host = "127.0.0.1"\n  user = "root"\n  password=""\n  port = 3306\n`,
+			`host = "127.0.0.1"\n  user = "root"\n  password=******\n  port = 3306\n`,
+		},
+	}
+	for i, str := range strs {
+		t.Logf("case #%d\n", i)
+		require.Equal(t, str.new, common.HideSensitive(str.old))
+	}
+}
