+++ br/pkg/lightning/importer/import.go
@@ -31,7 +31,6 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/kvproto/pkg/metapb"
-	"github.com/pingcap/tidb/br/pkg/backup"
 	berrors "github.com/pingcap/tidb/br/pkg/errors"
 	"github.com/pingcap/tidb/br/pkg/lightning/backend"
 	"github.com/pingcap/tidb/br/pkg/lightning/backend/encode"
@@ -53,6 +52,7 @@ import (
 	"github.com/pingcap/tidb/br/pkg/version"
 	"github.com/pingcap/tidb/br/pkg/version/build"
 	tidbconfig "github.com/pingcap/tidb/pkg/config"
+	"github.com/pingcap/tidb/pkg/distsql"
 	"github.com/pingcap/tidb/pkg/keyspace"
 	tidbkv "github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/meta/autoid"
@@ -1453,7 +1453,7 @@ func (rc *Controller) buildTablesRanges() []tidbkv.KeyRange {
 	var keyRanges []tidbkv.KeyRange
 	for _, dbInfo := range rc.dbInfos {
 		for _, tableInfo := range dbInfo.Tables {
-			if ranges, err := backup.BuildTableRanges(tableInfo.Core); err == nil {
+			if ranges, err := distsql.BuildTableRanges(tableInfo.Core); err == nil {
 				keyRanges = append(keyRanges, ranges...)
 			}
 		}
