+++ br/pkg/restore/pipeline_items.go
@@ -15,8 +15,8 @@ import (
 	"github.com/pingcap/tidb/br/pkg/metautil"
 	"github.com/pingcap/tidb/br/pkg/rtree"
 	"github.com/pingcap/tidb/br/pkg/summary"
-	"github.com/pingcap/tidb/br/pkg/utils"
 	"github.com/pingcap/tidb/pkg/parser/model"
+	"github.com/pingcap/tidb/pkg/util"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 )
@@ -328,7 +328,7 @@ func (b *tikvSender) splitWorker(ctx context.Context,
 		summary.CollectDuration("split region", elapsed)
 	}()
 
-	pool := utils.NewWorkerPool(concurrency, "split")
+	pool := util.NewWorkerPool(concurrency, "split")
 	for {
 		select {
 		case <-ectx.Done():
