+++ br/pkg/restore/split.go
@@ -22,6 +22,7 @@ import (
 	"github.com/pingcap/tidb/br/pkg/utils"
 	"github.com/pingcap/tidb/br/pkg/utils/iter"
 	"github.com/pingcap/tidb/pkg/tablecodec"
+	"github.com/pingcap/tidb/pkg/util"
 	"github.com/pingcap/tidb/pkg/util/codec"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
@@ -177,7 +178,7 @@ func (rs *RegionSplitter) executeSplitByKeys(
 		for _, region := range regions {
 			regionMap[region.Region.GetId()] = region
 		}
-		workerPool := utils.NewWorkerPool(uint(splitContext.storeCount)+1, "split keys")
+		workerPool := util.NewWorkerPool(uint(splitContext.storeCount)+1, "split keys")
 		eg, ectx := errgroup.WithContext(ctx)
 		for regionID, splitKeys := range splitKeyMap {
 			region := regionMap[regionID]
@@ -309,7 +310,7 @@ type LogSplitHelper struct {
 	tableSplitter map[int64]*split.SplitHelper
 	rules         map[int64]*RewriteRules
 	client        split.SplitClient
-	pool          *utils.WorkerPool
+	pool          *util.WorkerPool
 	eg            *errgroup.Group
 	regionsCh     chan []*split.RegionInfo
 
@@ -322,7 +323,7 @@ func NewLogSplitHelper(rules map[int64]*RewriteRules, client split.SplitClient, 
 		tableSplitter: make(map[int64]*split.SplitHelper),
 		rules:         rules,
 		client:        client,
-		pool:          utils.NewWorkerPool(128, "split region"),
+		pool:          util.NewWorkerPool(128, "split region"),
 		eg:            nil,
 
 		splitThreSholdSize: splitSize,
