+++ pkg/util/mathutil/math.go
@@ -98,3 +98,16 @@ func Clamp[T constraints.Ordered](n, min, max T) T {
 	}
 	return n
 }
+
+// NextPowerOfTwo returns the smallest power of two greater than or equal to `i`
+// Caller should guarantee that i > 0 and the return value is not overflow.
+func NextPowerOfTwo(i int64) int64 {
+	if i&(i-1) == 0 {
+		return i
+	}
+	i *= 2
+	for i&(i-1) != 0 {
+		i &= i - 1
+	}
+	return i
+}
